Moved duplicateProject and semverBumpPackageJson to plumbing
