+++ packages/@haiku/sdk-client/src/createProjectFiles.ts
@@ -10,10 +10,8 @@ import {
     fetchProjectConfigInfo,
     getAuthorNameOrFallback,
     getCopyrightNotice,
-    getDefaultIllustratorAssetPath,
     getHaikuCoreVersion,
     getOrganizationNameOrFallback,
-    getDefaultSketchAssetPath,
     getProjectNameLowerCase,
     getProjectNameSafeShort,
     getReactProjectName,
